tests refactor complete
